📑 ::상절점 컴포넌트 서스펜스 적용
