+++ src/store/point/pointStore.ts
@@ -1,0 +1,6 @@
+import { atom } from "recoil";
+
+export const pointViewTypeAtom = atom({
+  key: "point/isDormitoryPointView",
+  default: true,
+});
